fix: Set Number to Finite in Schemas & test replays.get in easy client
`S.Finite` is better because when generating a mock replay via
`FastCheck.sample()` it was generating `Infinity` & `-Infinity` which
were not being stringified into JSON correctly by nock. If nock wasn't
able to do it, then it's likely the same for the actual API assuming
that JSON doesn't support infinite numbers. It also doesn't seem
possible to have properties with infinite values since most would
require a replay that went on for forever etc.